+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UserDto {
-    private final int id;
+    private final Long id;
     private final String name;
     private final String email;
 }
